Renames v2 appointments transformer
